fixed more compiler niggles

+++ CCamera.cpp
@@ -25,7 +25,7 @@ ray3f CCamera::getRay(const vec2i& pos)
     if (m_dirty)
         computePlane();
 
-    vec3f planeCoord = (m_vecPerPixelX * pos.X) + (m_vecPerPixelY * (m_imageSize.Y - pos.Y));
+    vec3f planeCoord = (m_vecPerPixelX * (f32)pos.X) + (m_vecPerPixelY * (f32)(m_imageSize.Y - pos.Y));
     vec3f worldCoord = m_viewPlaneCorner + planeCoord;
     vec3f dir = worldCoord - m_eye;
     return ray3f(m_eye, dir);
@@ -111,8 +111,8 @@ void CCamera::computePlane()
     //m_viewPlaneCorner = m_lookAt - (u*viewPlaneW) - (v*viewPlaneH);
     m_viewPlaneCorner = (m_eye + viewDir) - (u*viewPlaneW) - (v*viewPlaneH);
 
-    m_vecPerPixelX = (u*viewPlaneW * 2) / m_imageSize.X;
-    m_vecPerPixelY = (v*viewPlaneH * 2) / m_imageSize.Y;
+    m_vecPerPixelX = (u*viewPlaneW * 2) / (f32)m_imageSize.X;
+    m_vecPerPixelY = (v*viewPlaneH * 2) / (f32)m_imageSize.Y;
     m_dirty = false;
 }
